requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
argcomplete==3.5.2
attrs==24.3.0
autoflake==2.3.1
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
datamodel-code-generator==0.26.4
dnspython==2.7.0
email_validator==2.2.0
Faker==33.1.0
genson==1.3.0
idna==3.10
inflect==5.6.2
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.5
MarkupSafe==3.0.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pydantic==2.10.5
pydantic_core==2.27.2
pyflakes==3.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
ruff==0.8.4
six==1.17.0
typing_extensions==4.12.2
urllib3==2.3.0
testit-adapter-pytest==3.5.5.post530
testit-api-client==6.0.1.post530
testit-python-commons==3.5.5.post530